yolo/requirements.txt
opencv-python
python-dotenv
torch
torchvision
aiohttp